refactor: converge prepare and source transformation boundaries
Split prepare into explicit transform, enrichment, and model modules while keeping the public orchestration surface stable. Also thin source iteration into a traversal facade over source-walk helpers without breaking the existing patch/import contract.
